Logement_Eco/requirements.txt
db-sqlite3==0.0.1
fastapi==0.115.4
httpx==0.28.0
Jinja2==3.1.4
pandas==2.2.3
pydantic==2.9.2
pyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
sqlalchemy==2.0.36
uvicorn==0.32.0